Compiled CleanOnionExample.L01.Abstractions

+++ src/Common/.Abstractions/Collections/Paged/PagedListOptions.cs
@@ -1,12 +1,12 @@
 ﻿namespace Common.Collections.Paged;
 
 public class PagedListOptions : IPagedListOptions {
-  public PagedListOptions(int pageNumber, int? pageSize, int totalItemCount) {
+  public PagedListOptions(int pageNumber, int? pageSize, long totalItemCount) {
     PageNumber = pageNumber;
     PageSize = pageSize;
     TotalItemCount = totalItemCount;
   }
-  public int TotalItemCount { get; }
+  public long TotalItemCount { get; }
   public int PageNumber { get; }
   public int? PageSize { get; }
 }
